Draft public release notes
This repository is for public release notes.

Initially this will hold a single file with release notes for the Clinician Portal. We can add files for other codebases as needed, but we technically already have public release notes for the mboiel app in the iOS and Google Play stores.

The intent is that we align releases in our private development repository with release notes here, so that once we’ve deployed a release to production, we can copy the text here to make it available to users outside the oVRcome organisation.

After the first 1.0.0 release, subsequent release notes will have sections `Added`, `FIxed`, `Changed`, and/or `Removed` as needed.